Wrote map file, more board setup
Set up a few variables in the board and made it so that moveable pieces have two 'states' - in a room or on the board. If on the board, use coordinates. If in a room, it'll be displayed under that room's list of contents. Also manually wrote the cluedo map into ASCII using a text editor.